requirements.txt
torch==1.13.1
torchvision==0.14.1
basicsr==1.4.2
diffusers==0.12.1
einops==0.6.0
k_diffusion==0.0.12
lark==1.1.5
numpy==1.24.1
Pillow==9.4.0
safetensors==0.2.8
tqdm==4.64.1
transformers==4.26.0
simple_websocket_server==0.4.2
websocket_client==1.5.0
bson==0.5.10
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0